Move misc. source path methods from WorkFiles to new mixin `NewspaperWorks::Data::PathHelper`
