+++ lib/newspaper_works/data.rb
@@ -1,4 +1,5 @@
 require 'newspaper_works/data/fileset_helper'
+require 'newspaper_works/data/path_helper'
 require 'newspaper_works/data/work_derivatives'
 require 'newspaper_works/data/work_files'
 require 'newspaper_works/data/work_file'
